probe.gl based logging
