+++ src/test-utils/package.json
@@ -38,12 +38,10 @@
     "bench-browser": "webpack-dev-server --env.bench --progress --hot --open",
     "test-rendering": "(cd test/render && webpack-dev-server --config webpack.config.test-rendering.js --progress --hot --open)"
   },
-  "dependencies": {
-    "probe.gl": "^0.2.2"
-  },
   "peerDependencies": {
     "deck.gl": ">=5.1.1",
-    "luma.gl": ">=5.1.0"
+    "luma.gl": ">=5.1.0",
+    "probe.gl": ">=1.0.0-alpha.1"
   },
   "devDependencies": {
     "babel-cli": "^6.22.2",
